+++ tests/test_errors.py
@@ -102,3 +102,24 @@ def test_nested_deserialization_error():
         match=r"required element '/test_model\[1\]/nested_model1\[1\]' not found"
     ):
         pb.from_xml(TestModel, xml)
+
+
+def test_order():
+    with pytest.raises(AssertionError, match=r"order element 'element2' not declared in model"):
+        @pb.model(order=('element1', 'element2'))
+        class TestModel:
+            element1 = pb.attr()
+
+
+def test_indexes_deserialization():
+
+    @pb.model()
+    class TestModel:
+        field = pb.field(idx=2)
+
+    obj = TestModel(field="value1")
+    with pytest.raises(
+            exc.SerializationError,
+            match=r"serialization can't be completed because field\[2\] is going to be serialized, "
+                  r"but field\[1\] is not serialized."):
+        pb.to_xml(obj)