+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+0.2.0 (2019-08-25)
+------------------
+
+- documentation added.
+- pre-commit hooks added.
+- model serialization order implemented.
+- private class attributes fixed.
+- indexed fields serialization fixed.
+
+
 0.1.0 (2019-08-13)
 ------------------